requirements.txt
Django==4.2.11
celery==4.4.7
vine==1.3.0
redis
django-celery-results
django-import-export==3.3.7
djangorestframework==3.15.1
PyJWT==2.8.0
requests==2.31.0
tzdata==2024.1
eventlet
